pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bbs2"
version = "0.1.0"
description = "Modular ham radio BBS for Linux AX.25 / KISS / NetROM"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # Async networking
    "pyserial-asyncio>=0.6",
    # Web interface
    "flask>=3.0",
    "flask-socketio>=5.3",
    # Database
    "aiosqlite>=0.19",
    # Auth
    "bcrypt>=4.1",
    # Config
    "pyyaml>=6.0",
    # Production WSGI (used with Flask-SocketIO threading mode)
    "simple-websocket>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
bbs2 = "bbs.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bbs*", "server*"]

[tool.setuptools.package-data]
"bbs" = ["py.typed"]
